+++ lib/utils/interceptor.dart
@@ -3,16 +3,22 @@ import 'dart:developer';
 import 'package:dio/dio.dart';
 import 'package:dio_shaker_interceptor/features/list_curl/data/curl_model.dart';
 import 'package:dio_shaker_interceptor/utils/curl_logger.dart';
+import 'package:dio_shaker_interceptor/utils/dsi_constants.dart';
 import 'package:dio_shaker_interceptor/utils/utils.dart';
 import 'package:flutter/material.dart';
 
 class DioShakerInterceptor extends InterceptorsWrapper {
   @override
   void onRequest(RequestOptions options, RequestInterceptorHandler handler) {
-    final curlModel = CurlModel(
+    final String id = generateDsiRequestId();
+    options.extra[kDsiRequestIdExtraKey] = id;
+
+    final CurlModel curlModel = CurlModel(
+      id: id,
       url: options.uri,
-      headers: options.headers,
+      headers: Map<String, dynamic>.from(options.headers),
       body: options.data,
+      requestSize: Utils.estimateSize(options.data),
       extra: options.extra,
       method: options.method,
       queryParameters: options.queryParameters,
@@ -24,21 +30,22 @@ class DioShakerInterceptor extends InterceptorsWrapper {
 
   @override
   void onResponse(Response response, ResponseInterceptorHandler handler) {
-    final curlModel = CurlLogs.instance.getCurlModelByUri(
-      response.requestOptions.uri,
-    );
+    final String? id = response.requestOptions.extra[kDsiRequestIdExtraKey] as String?;
+    if (id == null) {
+      super.onResponse(response, handler);
+      return;
+    }
+    final CurlModel? curlModel = CurlLogs.instance.getById(id);
     if (curlModel != null) {
       try {
-        final updatedCurlModel = curlModel.copyWith(
+        final CurlModel updatedCurlModel = curlModel.copyWith(
           status: response.statusCode,
           response: response.data,
+          responseSize: Utils.estimateSize(response.data),
         );
-        CurlLogs.instance.replace(
-          replace: curlModel,
-          save: updatedCurlModel,
-        );
+        CurlLogs.instance.replace(replace: curlModel, save: updatedCurlModel);
       } catch (e) {
-        debugPrint("DioshakerInterceptor error: $e");
+        debugPrint('DioshakerInterceptor error: $e');
       }
     }
     super.onResponse(response, handler);
@@ -46,22 +53,19 @@ class DioShakerInterceptor extends InterceptorsWrapper {
 
   @override
   void onError(DioException err, ErrorInterceptorHandler handler) {
-    if (err.response != null) {
-      final curlModel = CurlLogs.instance.getCurlModelByUri(
-        err.requestOptions.uri,
-      );
+    final String? id = err.requestOptions.extra[kDsiRequestIdExtraKey] as String?;
+    if (id != null && err.response != null) {
+      final CurlModel? curlModel = CurlLogs.instance.getById(id);
       if (curlModel != null) {
         try {
-          final updatedCurlModel = curlModel.copyWith(
+          final CurlModel updatedCurlModel = curlModel.copyWith(
             status: err.response?.statusCode,
             response: err.response?.data,
+            responseSize: Utils.estimateSize(err.response?.data),
           );
-          CurlLogs.instance.replace(
-            replace: curlModel,
-            save: updatedCurlModel,
-          );
+          CurlLogs.instance.replace(replace: curlModel, save: updatedCurlModel);
         } catch (e) {
-          debugPrint("DioshakerInterceptor error: $e");
+          debugPrint('DioshakerInterceptor error: $e');
         }
       }
     }
@@ -72,7 +76,7 @@ class DioShakerInterceptor extends InterceptorsWrapper {
 class CurlInterceptor extends InterceptorsWrapper {
   @override
   void onRequest(RequestOptions options, RequestInterceptorHandler handler) {
-    final curlModel = CurlModel(
+    final CurlModel curlModel = CurlModel(
       url: options.uri,
       headers: options.headers,
       body: options.data,